[IOTDB-1081] New TsFile Format (V3) (#2184)

+++ server/src/main/java/org/apache/iotdb/db/query/externalsort/serialize/impl/FixLengthIExternalSortFileDeserializer.java
@@ -83,7 +83,7 @@ public class FixLengthIExternalSortFileDeserializer implements IExternalSortFile
   }
 
   private TSDataType readHeader() throws IOException {
-    return TSDataType.deserialize(ReadWriteIOUtils.readShort(inputStream));
+    return TSDataType.deserialize(ReadWriteIOUtils.readByte(inputStream));
   }
 
   private void setReader(TSDataType type) {
